cleanups and lambda full migration
