+++ .github/workflows/deploy-bots.yaml
@@ -26,7 +26,6 @@ on:
 
 jobs:
   terraform-infra:
-    # if: startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main'
     runs-on: ubuntu-latest
     name: Deploy Infra to AWS
     steps:
@@ -127,52 +126,7 @@ jobs:
     outputs:
       json: ${{ steps.meta.outputs.json }}
 
-  build-and-push-oracles-monitor-image:
-    needs: [terraform-infra]
-    env:
-      REGISTRY: ghcr.io
-      IMAGE_NAME: midas-protocol/oracles-monitor
-
-    runs-on: ubuntu-latest
-    permissions:
-      contents: read
-      packages: write
-
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          submodules: recursive
-          token: '${{ secrets.ACCESS_TOKEN }}'
-
-      - name: Log in to the Container registry
-        uses: docker/login-action@f054a8b539a109f9f41c372932f1ae047eff08c9
-        with:
-          registry: ${{ env.REGISTRY }}
-          username: ${{ github.actor }}
-          password: ${{ secrets.GITHUB_TOKEN }}
-
-      - name: Extract metadata (tags, labels) for Docker
-        id: meta
-        uses: docker/metadata-action@98669ae865ea3cffbcbaa878cf57c20bbf1c6c38
-        with:
-          images: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}
-          tags: |
-            type=sha,format=long
-
-      - name: Build and push Docker image
-        uses: docker/build-push-action@ad44023a93711e3deb337508980b4b5e9bcdc5dc
-        with:
-          context: .
-          push: true
-          tags: ${{ steps.meta.outputs.tags }}
-          labels: ${{ steps.meta.outputs.labels }}
-          file: docker/oracles-monitor/Dockerfile
-
-    outputs:
-      json: ${{ steps.meta.outputs.json }}
-
   build-and-push-oracles-monitor-ecr:
-    # if: github.ref == 'refs/heads/main' || github.ref == 'refs/heads/staging' || github.ref == 'refs/heads/prod'
     needs: [terraform-infra]
     runs-on: ubuntu-latest
     env:
@@ -214,19 +168,13 @@ jobs:
 
       - name: Build, tag, and push docker image to Amazon ECR Public
         run: |
-          docker build -f docker/oracles-monitor/Dockerfile.lambda -t $REGISTRY/$REPOSITORY:$IMAGE_TAG -t ghcr.io/midas-protocol/oracles-monitor:$IMAGE_TAG .
+          docker build -f docker/oracles-monitor/Dockerfile -t $REGISTRY/$REPOSITORY:$IMAGE_TAG -t ghcr.io/midas-protocol/oracles-monitor:$IMAGE_TAG .
           docker push $REGISTRY/$REPOSITORY:$IMAGE_TAG
           docker push ghcr.io/midas-protocol/oracles-monitor:$IMAGE_TAG
 
   smoke-tests:
     runs-on: ubuntu-latest
-    needs:
-      [
-        build-and-push-liquidator-image,
-        build-and-push-oracles-monitor-image,
-        build-and-push-oracles-monitor-ecr,
-        lint-bots,
-      ]
+    needs: [build-and-push-liquidator-image, build-and-push-oracles-monitor-ecr, lint-bots]
     env:
       ETHEREUM_ADMIN_ACCOUNT: '${{ secrets.ETHEREUM_ADMIN_ACCOUNT }}'
       ETHEREUM_ADMIN_PRIVATE_KEY: '${{ secrets.ETHEREUM_ADMIN_PRIVATE_KEY }}'
