+++ ops/prod/moonbeam.tf
@@ -21,6 +21,34 @@ module "moonbeam_mainnet_oracle_price_change_verifier" {
 }
 
 
+module "monbeam_mainnet_oracle_feed_verifier" {
+  source              = "../modules/lambda"
+  ecr_repository_name = "oracles-monitor"
+  docker_image_tag    = var.oracles_monitor_image_tag
+  container_family    = "feed-verifier"
+  environment         = "mainnet"
+  chain_id            = local.monbeam_mainnet_chain_id
+  container_env_vars = merge(
+    local.oracle_feed_verifier_lambda_variables,
+    { WEB3_HTTP_PROVIDER_URL = local.monbeam_mainnet_rpc_1 }
+  )
+  schedule_expression = "rate(1 hour)"
+}
+
+module "monbeam_mainnet_oracle_price_verifier" {
+  source              = "../modules/lambda"
+  ecr_repository_name = "oracles-monitor"
+  docker_image_tag    = var.oracles_monitor_image_tag
+  container_family    = "price-verifier"
+  environment         = "mainnet"
+  chain_id            = local.monbeam_mainnet_chain_id
+  container_env_vars = merge(
+    local.oracle_price_verifier_lambda_variables,
+    { WEB3_HTTP_PROVIDER_URL = local.monbeam_mainnet_rpc_1 }
+  )
+  schedule_expression = "rate(5 minutes)"
+}
+
 module "moonbeam_mainnet_liquidation_cron" {
   source                  = "../modules/cron"
   service_security_groups = module.network.ecs_task_sg
