+++ packages/monitors/oracle/src/services/verifiers/feed/providers/chainlink.ts
@@ -1,6 +1,6 @@
 import { Contract } from "ethers";
 
-import { logger } from "../../../../index";
+import { logger } from "../../../../logger";
 import { FeedVerifierConfig, InvalidReason, PriceFeedValidity, VerifyFeedParams } from "../../../../types";
 
 export async function verifyChainLinkOraclePriceFeed(
