+++ packages/monitors/oracle/tests/run.ts
@@ -1,12 +1,20 @@
 import { JsonRpcProvider } from "@ethersproject/providers";
+import { MidasSdk } from "@midas-capital/sdk";
 import { SupportedChains } from "@midas-capital/types";
 import { Wallet } from "ethers";
 
-import { setUpSdk } from "../src";
-import { assets } from "../src/config";
+import { assets, configs, verifiers } from "../src/config";
 import { baseConfig } from "../src/config/variables";
-import { runVerifier } from "../src/run";
-import { Services } from "../src/types";
+import { logger, setUpSdk } from "../src/logger";
+import { BatchVerifier } from "../src/services/verifier";
+import { OracleVerifierAsset, Services } from "../src/types";
+
+export async function runVerifier(sdk: MidasSdk, service: Services, assetsOverride?: OracleVerifierAsset[]) {
+  logger.info(`RUNNING SERVICE: ${service}`);
+  const assetsToVerify = assetsOverride ? assetsOverride : assets[service];
+  const verifier = new BatchVerifier(sdk, assetsToVerify);
+  await verifier.batchVerify(verifiers[service], configs[service]);
+}
 
 (async function () {
   const chainId: number = process.env.TARGET_CHAIN_ID ? parseInt(process.env.TARGET_CHAIN_ID) : SupportedChains.ganache;
@@ -16,7 +24,15 @@ import { Services } from "../src/types";
 
   const assetsOverride = assets[Services.FeedVerifier].slice(0, 1);
   const priceChangeVerifierAssetsOverride = assets[Services.PriceChangeVerifier].slice(0, 1);
-  runVerifier(midasSdk, Services.FeedVerifier, assetsOverride);
-  runVerifier(midasSdk, Services.PriceVerifier, assetsOverride);
-  runVerifier(midasSdk, Services.PriceChangeVerifier, priceChangeVerifierAssetsOverride);
+
+  switch (baseConfig.service) {
+    case Services.FeedVerifier:
+      runVerifier(midasSdk, Services.FeedVerifier, assetsOverride);
+      break;
+    case Services.PriceVerifier:
+      runVerifier(midasSdk, Services.PriceVerifier, assetsOverride);
+      break;
+    case Services.PriceChangeVerifier:
+      runVerifier(midasSdk, Services.PriceChangeVerifier, priceChangeVerifierAssetsOverride);
+  }
 })();
